feat: replace authorizaton validation  in controllers by Gates (#8)
* feat: implement authorization for club actions using Gate and add ClubPolicy

* feat: implement court authorization using Gate and add CourtPolicy

* chore: fix code styling

* refactor: clean up unused Gate imports and improve code formatting in controllers

* refactor: clean up unused Gate imports and improve code formatting in controllers

* refactor: update query mock to use QueryExecuted instead of Builder in AppServiceProviderTest

* refactor: update query mock to use QueryExecuted instead of Builder in AppServiceProviderTest

* refactor: simplify mock declaration for QueryExecuted in AppServiceProviderTest

* refactor: optimize test execution by adjusting Pest command parameters

* refactor: optimize test execution by adjusting Pest command parameters

* refactor: enhance test performance by adjusting Pest parameters and configuring Rector for parallel processing

* refactor: switch database connection from MySQL to SQLite for testing

+++ app/Http/Controllers/Club/CourtPriceRuleController.php
@@ -9,8 +9,8 @@ use App\Http\Resources\Club\ShowCourtPriceRuleResource;
 use App\Models\Club;
 use App\Models\Court;
 use App\Services\CourtPriceRulesShowBuilderService;
-use App\Services\OwnershipVerifierService;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 use Throwable;
 
@@ -22,11 +22,10 @@ final class CourtPriceRuleController
     public function store(
         StoreCourtPriceRuleRequest $request,
         Club $club,
-        Court $court,
-        OwnershipVerifierService $ownershipVerifier,
+        Court $court
     ): Response {
-        $this->ensureCourtBelongsToClub($club, $court);
-        $ownershipVerifier->handle($court->club);
+
+        Gate::authorize('update', [$court, $club]);
 
         DB::transaction(function () use ($court, $request): void {
             $court->priceRules()->delete();
@@ -54,21 +53,12 @@ final class CourtPriceRuleController
     public function show(
         Club $club,
         Court $court,
-        OwnershipVerifierService $ownershipVerifier,
         CourtPriceRulesShowBuilderService $courtPriceRulesShowBuilder,
     ): ShowCourtPriceRuleResource {
-        $this->ensureCourtBelongsToClub($club, $court);
-        $ownershipVerifier->handle($court->club);
+        Gate::authorize('view', [$court, $club]);
 
         return new ShowCourtPriceRuleResource(
             $courtPriceRulesShowBuilder->handle($court),
         );
-    }
-
-    private function ensureCourtBelongsToClub(Club $club, Court $court): void
-    {
-        if ($court->club_id !== $club->id) {
-            abort(404, __('validation.resource_not_found'));
-        }
     }
 }
